Fixed two issues in AG raised by Thomas van Noort:
Patterns for nested tuples are allowed again, such as
loc.(a,(b,c)) = f x
These were given up in rev.145, but are needed for compiling Generic Haskell.

In literate mode, code blocks are started with \begin{Code}.
The capitalized Code is strange for Haskell programmers, because in Haskell \begin{code} is used. 
For compatibility with the Haskell convention, we now allow this form as well.


git-svn-id: https://svn.science.uu.nl/repos/project.STEC.uuagc/uuagc/trunk@167 30b93d28-2606-0410-b7f1-95e0edb40354