'Problem': Large AG modules (like MainAG of EHC) result in big Haskell files. This has some downsides:   * GHC runs out of memory during compilation with optimisations on   * Large compilation times
Idea:
Split the big Haskell file in smaller files. One would think that this 
is difficult because traversal functions are typically mutually 
recursive and mutually recursive functions are nasty to separate into 
different modules. Fortunately, the semantic functions of an AG are not 
recursive (the recursive part is in the folds), so we can generate a 
separate module for semantic functions.

Downsides:
  * Lots of small modules
  * Possibly poorer inlining by the optimizer
  * Produces incorrect code when code blocks in the AG file refer to 
semantic functions (for example, with higher-order attributes)

This is now available as a feature, just add the
  --sepsemmods
commandline option. For an AG module <Name>, it generates the modules:

<Name>.hs
<Name>_common.hs
<Name>_NT1.hs
...
<Name>_NTk.hs

Todo:
  * At the moment there is a restriction that no module headers may be 
specified in the AG file. Have to figure out how to deal with those.
  * Add the option to group 'small' sets of semantic functions into a 
single file (to prevent generation of a lot of very tiny modules). Or 
analyze it automatically with some heuristic for the code-size of a 
semantic function.


git-svn-id: https://svn.science.uu.nl/repos/project.STEC.uuagc/uuagc/trunk@212 30b93d28-2606-0410-b7f1-95e0edb40354