Added an option:
> PRAGMA genusetraces
(or --genusetraces on the commandline)

It adds a |trace| statement around attribute occurrences on the RHS.

Due to the RHS basically being strings of haskell, actually adding these 
traces is not as easy as it should be. Therefore, at the moment, traces 
are not generated for: copy rules, default rules, collection rules, and 
access to terminal values. In other words: only for those |@field.attr| 
occurrences in your code. 


git-svn-id: https://svn.science.uu.nl/repos/project.STEC.uuagc/uuagc/trunk@272 30b93d28-2606-0410-b7f1-95e0edb40354