Conservatively tweaked the implementation of --sepsemmods a bit. This option compiled a big AG to many small haskell modules.
It is now possible to call a haskell function from inside a semantic function, that itself calls a wrapper function. This was not 
possible when the --sepsemmods flag was present, because some functions (wrappers etc) were not in scope. Now it is possible, 
because but you may need to add additional import statements, and arrange imports in such a way to prevent cyclic imports.

Additional notation:

Aside from writing:

imports { ... haskell code ... }
optpragma { ... haskell code ... }

The following is now also allowed:

ATTACH Nonterm { ... haskell code ... }

which prints the haskell code just below the code of the semantic function of that nonterminal. No real effect when the 
--sepsemmods flag is absent. When the flag is present, it allowed you to put certain haskell blocks in a semantic-function module 
instead of the 'common' module. The keyword ATTACH is required to prevent a parsing ambiguity with uppercase identifiers.


git-svn-id: https://svn.science.uu.nl/repos/project.STEC.uuagc/uuagc/trunk@285 30b93d28-2606-0410-b7f1-95e0edb40354