The SEPSEMMODS pragma now has its original semantics back.
With this pragma, the generated haskell code is split over multiple 
haskell files. This is sometimes not possible when you call wrappers 
from the semantic functions, either indirectly through haskell or by 
using a higher-order attribute. These wrappers are defined in a seperate 
module that imports the individual modules with semantic functions. 
Trying to import that module leads to cyclic module dependencies. I 
experimented a bit with it (hoping that I could shuffle around the code 
in such a way that the ag programmer can specify additional imports such 
that the generated code compiles without having cycles), but I'm afraid 
this is not possible without analyzing the Haskell blocks (which is not 
something we want to do for the AG system).

Conclusion: the SEPSEMMODS feature and using some wrappers/higher-order 
attributes do not go along together. Too bad for some of the Proxima 
people, they'll just have to use faster machines with more memory to 
compile some of the huge Haskell files.


git-svn-id: https://svn.science.uu.nl/repos/project.STEC.uuagc/uuagc/trunk@286 30b93d28-2606-0410-b7f1-95e0edb40354