Allowed a (higher-order) inst-attribute to shadow/transform a child of a production.
Abstract example:

> -- a production C with a child k of nonterm D
> DATA D | C          k : D
>
> -- we extend k : D such that it will become k : D_k
> DATA D_k | Extend   k : D
>
> SEM D
>   | C   inst.k : D_k  -- note that a child k already exists
>
>         -- the code for this virtual 'k' is a function that
>         -- takes the original semantics of k (T_D) and must
>         -- provide a semantics of D_k (i.e. of type T_D_k)
>         inst.k = \semD -> sem_D_k_Extend semD
>
> -- declare attributes on D and D_k
> -- define attributes

In other words: when defining an inst-attribute, if it shadows
an existing child, its definition must be a function that
takes the semantis of the shadowed child. If not, it is just
an ordinary value.

You can use this mechanism to define attributes of children in
isolation, conditionally dependent on other attributes, orthogonal
to copy rules.


An a more concrete example, we look at the threading of
environments for a recursive or non-recursive let
(indicated by a boolean).

> DATA Expr
>   | Let  isRec : {Bool}
>          nm    : {String}
>          def   : Expr
>          body  : Expr

Instead of having "def" to be an Expr, we at
runtime put a type in between to make a
difference between a recursive or non-recursive
def.

> DATA ExprLetDef
>   | Rec     nm : String  expr : Expr
>   | NonRec  nm : String  expr : Expr
>
> SET ExprShared = Expr ExprLetDef
> ATTR ExprShared [ env : Env
                  | subst : Subst
                  | errs USE {`mappend`} {mempty} : Errs ]

The ExprLetDef has an additional attribute that
gives the environment to be used for the body
of the Let.

> ATTR ExprLetDef [ | | env : Env ]

We can now simply define the code for
the Let as choosing between a rec or
nonrec production on top of the "def" child.

> SEM Expr
>   | Let  inst.def : ExprLetDef
>          inst.def = if @isRec
>                     then sem_ExprLetDef_Rec @nm
>                     else sem_ExprLetDef_NonRec @nm

The behavior for the environment, we can now write
seperately, in isolation from the rest of the code
for the Let.

> SEM ExprLetDef
>   | Rec NonRec
>          +lhs.env = Map.insert @nm @expr.ty
> 
>   | Rec  loc.freshTy : UNIQUEREF subst
>          expr.env = Map.insert @nm @loc.freshTy @lhs.env

Suppose we may also have a type signature for a binding.

> DATA Expr
>  | Let  mbSig : {Maybe Sig}

We can then add a check that the signature matches the
inferred type, depending if this signature is indeed
available.

DATA ExprLetDef
  | CheckSig  sig : {Sig}  expr : ExprLetDef

SEM Expr
  | Let  AROUND def = maybe (const id) sem_ExprLetDef_CheckSig @mbSig

To check the signature, we need to convert the signature to some type,
and do some unification that involves the substitution and possibly
errors:

SEM ExprLetDef
  | CheckSig  loc.expect = sigToTy @sig
              (lhs.subst, lhs.errs) = unify @expr.ty @loc.expect @expr.subst

The above example is contrived in the sense that the
choice between which ExprLetDef-productions to insert
was directly based on syntax. Hence, we could also have
changed the grammar such that the parser inserts these
additional productions in between.

> DATA Expr | Let  def : ExprLetDef
> DATA ExprLetDef | Rec NonRec CheckSig

However, the difference is that you can now make this decision based
on the values of attributes. Question remains if this is really useful.


git-svn-id: https://svn.science.uu.nl/repos/project.STEC.uuagc/uuagc/trunk@362 30b93d28-2606-0410-b7f1-95e0edb40354