Minor change to the generation of visit functions. You are now no longer required to give type signatures to local attributes that are needed in more than one visit.
Instead of explicitly passing state from one visit to
the other via parameters and partial parametrization,
we define the function to be used for the next
visit inside the definition of the previous visit.

Positive effects: type signatures for locals not required anymore.
Negative effects: none, to the best of my knowledge.
The visit function of visit 2 and later are now not toplevel
functions anymore and cannot be called from user code.
This, however, was not allowed in the first place.

Compiled EH99 successfully with these changes, and will upload a new version to Hackage.


git-svn-id: https://svn.science.uu.nl/repos/project.STEC.uuagc/uuagc/trunk@365 30b93d28-2606-0410-b7f1-95e0edb40354