Implemented Atze's feature request: generation of data types using record syntax.
To enable, specify the commandline option: --datarecords
Or in your file: PRAGMA datarecords
This feature is also implied when you use the commandline option -a.

You then get for each symbol x :: ty of a production C of nonterminal N
a data type declaration:

> data N =
> ...
> | C { ...
>     , x_N_C :: ty
>     , ... }
> ...

This feature will be present the next time uuagc is released
to hackage.


git-svn-id: https://svn.science.uu.nl/repos/project.STEC.uuagc/uuagc/trunk@452 30b93d28-2606-0410-b7f1-95e0edb40354