Fixed a slight bug in the KW code generation which seemed only to occur after I turned off the 'multirule'. The problem was that the synthesized 'lhs' attrs computed by a visit where not put in the 'needed' intra-visit attr set, thus an lhs-synthesized attribute that is computed in an earlier visit than where its value is needed (which can happen if a rule defines multiple attrs) would not be carried over.
The multi-rule has as side effect that it prevented this
bug from showing up. It introduces a separate
extraction rule for each attribute that is part of a
rule that defines multiple attrs. Such an extraction rule
has no other dependencies and defines no other attributes,
thus an extraction rule for an lhs-attr is scheduled to the
visit that returns the value of that attr, and thus the
value is in scope in this case.

It is now fixed, and we can thus get rid of it for the
KW code generation. It was always a hack in my eyes to
put a plaster on a design error in the dependency graph
of the Kastens code generation code.

Anyway, in exactly the situation that I just fixed, it
is actually better not to use the multi-rule: it unnecessarily
requires that attributes computed by the same rule are
alive until the last of these attributes if delivered, which
is a waste when some of them are not needed anymore after
previous visits.


git-svn-id: https://svn.science.uu.nl/repos/project.STEC.uuagc/uuagc/trunk@487 30b93d28-2606-0410-b7f1-95e0edb40354