Made it possible to wrap error messages in rules, which will only be yielded if the rule is actually scheduled. (supported by KW code generation only)
During the DefaultRules.ag phase, either critical errors are reported,
or the AG is made statically well-formed. Any missing rule for an
inherited or synthesized attribute is automatically inserted with
"error ..." as right-hand side and a warning is generated.
What I changed is that for rules of productions of a nonterminal that
is marked as a wrapper, the warning message is wrapped in the
generated rule.

Our KW code generation schedules these rules only when really needed.
(technical detail: these error-rules do not reference any attributes
in their right-hand side, which ensures the above property even in
the presence of our more eager scheduling strategy).

This feature thus now fullfils a wish of Atze: you can now define a
big AG with many inherited and synthesized attributes, then define
a couple of additional nonterminals to be used as start symbols
(interfacing nonterminals) which only use a subset of the inh/syn
attributes. Mark these nonterminals as wrappers and copy rules do
the rest.


git-svn-id: https://svn.science.uu.nl/repos/project.STEC.uuagc/uuagc/trunk@490 30b93d28-2606-0410-b7f1-95e0edb40354