Added an option --latehigherorderbinding which adds an automatic inherited attribute which passes a dictionary of all semantic functions as inherited attribute.
With this feature you can use the sepsemmods feature in combination
with higher order attributes. This combination should make the compilation
fast(er) even when using the KW code generation.

Todo: the KW code generation doesn't work well with
"deforested" higher-order attributes. This is also not nicely
represented inside UUAG. That can be improved.
Mini-todo: in case of the non-forested higher-order attribute,
I should change the call to the sem function to automatically
use this dictionary. But first sleep.


git-svn-id: https://svn.science.uu.nl/repos/project.STEC.uuagc/uuagc/trunk@492 30b93d28-2606-0410-b7f1-95e0edb40354